# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/contact": {
      "post": {
        "operationId": "ContactController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create contact",
        "tags": [
          "Contact"
        ]
      }
    },
...
    "/api/faq/{id}": {
      "get": {
        "operationId": "FaqController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get faq by id",
        "tags": [
          "Faq"
        ]
      }
    },
    "/api/admin/faq": {
      "post": {
        "operationId": "FaqController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFaqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create faq",
        "tags": [
          "Faq"
        ]
      },
      "get": {
        "operationId": "FaqController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read all faqs",
        "tags": [
          "Faq"
        ]
      }
    },
    "/api/admin/faq/batch-create": {
      "post": {
        "operationId": "FaqController_batchCreate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreateFaqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Batch create or update faqs",
        "tags": [
          "Faq"
        ]
      }
    },
    "/api/admin/faq/{id}": {
      "get": {
        "operationId": "FaqController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read one faq",
        "tags": [
          "Faq"
        ]
      },
      "patch": {
        "operationId": "FaqController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFaqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update faq",
        "tags": [
          "Faq"
        ]
      },
      "delete": {
        "operationId": "FaqController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete faq",
        "tags": [
          "Faq"
        ]
      }
    },
    "/api/admin/contact": {
      "post": {
        "operationId": "ContactController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create contact",
        "tags": [
          "Contact"
        ]
      },
      "get": {
        "operationId": "ContactController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read all contacts",
        "tags": [
          "Contact"
        ]
      }
    },
    "/api/admin/contact/{id}": {
      "get": {
        "operationId": "ContactController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read one contact",
        "tags": [
          "Contact"
        ]
      },
      "patch": {
        "operationId": "ContactController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContactDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update contact",
        "tags": [
          "Contact"
        ]
      },
      "delete": {
        "operationId": "ContactController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete contact",
        "tags": [
          "Contact"
        ]
      }
    },
...
    "/api/admin/notification": {
      "get": {
        "operationId": "NotificationController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all notifications",
        "tags": [
          "Notification"
        ]
      },
      "delete": {
        "operationId": "NotificationController_removeAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete all notifications",
        "tags": [
          "Notification"
        ]
      }
    },
    "/api/admin/notification/{id}": {
      "delete": {
        "operationId": "NotificationController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete notification",
        "tags": [
          "Notification"
        ]
      }
    },
...
  },
  "info": {
    "title": "Application Name api",
    "description": "Application Name api docs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateContactDto": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string",
            "description": "First name",
            "example": "John"
          },
          "last_name": {
            "type": "string",
            "description": "Last name",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "description": "Email",
            "example": "john.doe@example.com"
          },
          "phone_number": {
            "type": "string",
            "description": "Phone number",
            "example": "+1234567890"
          },
          "message": {
            "type": "string",
            "description": "Message",
            "example": "Hello, I have a question about your product."
          }
        },
        "required": [
          "first_name",
          "last_name",
          "email",
          "phone_number",
          "message"
        ]
      },
      "CreateFaqDto": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string",
            "description": "FAQ question",
            "example": "What is the price of the hotel?"
          },
          "answer": {
            "type": "string",
            "description": "FAQ answer",
            "example": "The price of the hotel is 10000 BDT per night."
          },
          "sort_order": {
            "type": "number",
            "description": "Sort order",
            "example": 1
          }
        },
        "required": [
          "question",
          "answer",
          "sort_order"
        ]
      },
      "BatchCreateFaqDto": {
        "type": "object",
        "properties": {
          "faqs": {
            "description": "Faq data array",
            "example": [
              {
                "question": "What is the price of the hotel?",
                "answer": "The price of the hotel is 10000 BDT per night.",
                "sort_order": 1
              }
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "faqs"
        ]
      },
      "UpdateFaqDto": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string",
            "description": "FAQ question",
            "example": "What is the price of the hotel?"
          },
          "answer": {
            "type": "string",
            "description": "FAQ answer",
            "example": "The price of the hotel is 10000 BDT per night."
          },
          "sort_order": {
            "type": "number",
            "description": "Sort order",
            "example": 1
          }
        }
      },
      "UpdateContactDto": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string",
            "description": "First name",
            "example": "John"
          },
          "last_name": {
            "type": "string",
            "description": "Last name",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "description": "Email",
            "example": "john.doe@example.com"
          },
          "phone_number": {
            "type": "string",
            "description": "Phone number",
            "example": "+1234567890"
          },
          "message": {
            "type": "string",
            "description": "Message",
            "example": "Hello, I have a question about your product."
          }
        }
      },
...
    }
  }
}